Merged PR 5684: Merge 52640_create_Productivity3D_designs_endpoint to Dev
The main objective of this piece of work was to add the `api/v2/profiles/design` API. It returns a design profile slice for a given project and imported file.

This was an area of Productivity 3D Services and Raptor that was new to me. Please pay special attention to the formation of `design descriptors`, `conversion to Raptor data types` and use of the `ProfilesHelper.convertProductionDataProfileResult`. While I get a response back (attached), I don't know if it's correct or not.

Because of the work necessary to build out the framework classes for my SpecFlow acceptance tests I've submitted the PR early, with those tests coming in a later commit.

Outstanding questions:
1. What should be returned when composite or design profiles fail, e.g. Raptor returns an error code, currently we return `null`.