Merged PR 6275: Merge 57238_remove_Filter_property_layerType to Dev
Removed Filter::layerType from the Models package. This PR deals with the changes to Filter service to accomodate that change.

**Reproduce:**
Send a new filter request to the Filter svc with a layerNumber property in the filterJson, e.g.

```
{  "filterUid": "",
  "name": "Test bug 57238",
  "filterJson": "{\"startUTC\":\"2012-10-31T06:00:00Z\",\"endUTC\":\"2012-11-03T05:59:59Z\",\"layerNumber\":\"0\"}"}
}
```

The response should come back as `{"Code":2034,"Message":"Layer number error. To use the layer number filter, layer type must be specified"}`.

**After fix:**
With the fix applied the above request, including layerNumber attribute, produces a `success` response and the filter is created.

**Note the two tests I ignored** I need to talk to Elspeth when she returns Tuesday because I don't see how Filter ever validated WSG points when there wasn't a polygonUid or polygonName.