Merged PR 7307: Merge 62013-No-Data to Dev
I finished the work started by Elspeth so we return a null/empty response object if we receive 'no data' from Raptor for CMV, MDP, Passcount, Speed and Temperature.

Need minor re-work in the A/C test project to handle the possibility of null object responses.

Also put in a temporary fix to the A/C image diff class to fix a filepath too long error. This needs proper attention in the short term, as part of story #62188.

Related work items: #62013