Merged PR 7398: Merge 62533-Add-dateRangeName-to-filter-response to Dev
Updates to support the default serialization of `Filter::DateRangeName` during the egress of Filter objects.

When the filter response returns now, for Custom (or any dateRangeType) the dateRangeType and dateRangeName are included in the response, e.g.

```
{
    "filterDescriptor": {
        "filterUid": "9e4d8fd6-d7d6-464f-9af9-f87d18fa8d7f",
        "name": "",
        "filterJson": "{\"startUTC\":\"2017-12-12T06:00:00Z\",\"endUTC\":\"2017-12-17T05:59:59Z\",\"dateRangeType\":7,\"dateRangeName\":\"Custom\"}"
    },
    "Code": 0,
    "Message": "success"
}
```

Due to the handling of response JSON a number of acceptance tests needed enhancing to use valid JSON.

Related work items: #62533