Merged PR 7753: #64111 - Refactored Models::Filter object and updated Proxies package to use
The goal of this story is to provide a consistent API interface for the Gladiators to develop against. Specifically it means using regular lower camel case on response properties. The issue was first noted on the Filter object's **designUID** but during investigation a number of other affected properties were found.

The fix applies only to the Filter object and changed some of the JsonProperty PropertyName attribute values to use correct lower camel case. Also changed the actual property name to be upper camel case to align class properties with VSS code styling guidelines.

- The Models package minor version was updated to reflect the breaking change.
- The Proxies package needed to be re-published after being updated to reference the new Models package version.
- The Filter service and 3DP service projects needed to be updated to use new Models and Proxies packages. (The Filter service also needed the Repository package upgraded to fix changes made earlier on an unrelated story.)
- A number of tests were fixed to reflect the change in case for property names.

After making the changes I created a filter and received the following response:

Note, the lowerCamelCase of startUtc, endUtc and designUid.

```
{
    "filterDescriptor": {
        "filterUid": "04623c05-ad1d-46bc-a723-0a8f987f7880",
        "name": "57492-60-2",
        "filterJson": "{\"startUtc\":\"2012-10-30T00:12:09.109\",\"endUtc\":\"2017-08-17T00:03:41\",\"designUid\":\"dd64fe2e-6f27-4a78-82a3-0c0e8a5e84ff\",\"forwardDirection\":false}"
    },
    "Code": 0,
    "Message": "success"
}
```

Related work items: #64111