Merged PR 7961: Merge Fix_null_StationOffset_metrics to Dev
During development of #55084 we found the Station Offset report from 3DP Raptor services wasn't returning the statistic information for each station. This should contain the min, max and average values of each of the six fields. This bug fix ensure those values are added. The objects are only serialized if any of the offsets in the station contain valid values.

e.g. if the offset Elevation is valid, i.e. not null, and the query has selected to include elevation, then the min/max/avg objects will contain respective values for Elevation.

See attached .json file for an example. Three of the offsets don't contain CMV, PassCount or Temperature data, but because the -3 and 5 offsets do we have these three metrics included in the `minimum`, `maximum` and `average` objects.

[station-offset-wtih-min-max-avg-values.txt](https://visionlink.visualstudio.com/DefaultCollection/_apis/git/repositories/7fa2b347-dd6a-422c-97af-9b9e9ff8b430/pullRequests/7961/attachments/station-offset-wtih-min-max-avg-values.txt)