Merged PR 8548: Return early if elevation type is invalid, e.g. None.
## Problem ##
> As a developer I want the 3DP Raptor summary volumes to return early if incorrect data has been provided and the VolumeCalcType value is 'None'.
At the moment we're sending the request to Raptor but it's moot in this use case and could be handled more gracefully by the controller, returning a useful error message to the client.

## Impact ##
A large amount of work continues to occur despite the invalid request. This includes a Summary Volumes request to Raptor.

## Solution ##
Simply check and fail early if `ElevationType=None` for a Summary Volumes request.

The following benchmarks were performed on locally running 3DP service with Mock Web API backing services but remote Raptor service.
Configuration was x64 and Release build.

A warm-up run was performed for each run.

*(Results are in milliseconds)*

| Fix Applied | Iterations | Run #1 | Run #2  | Avg    | Requests/Minute |
| :-:         | :-:        | :-:    | :-:     | :----: | :---:           |
| No          | 10         | 43054  | 42549   | 42801  | 14.01           |
| No          | 100        | 429699 | 428643  | 429171 | 13.98           |
| Yes         | 10         | 6295   | 6246    | 6271   | 95.67           |
| Yes         | 100        | 65941  | 62760   | 64350  | 93.24           |

> A performance gain of 574.92%

The runs were performed using Postman Runner. There may be an overhead in the sending of requests, but as all tests were performed with the same query from Runner we can look at the results in relationship to one another without bias.

Related work items: #67112