Merged PR 8845: Fix invalidation of response cache objects (by project)
The first implementation of response caching didn't invalidate the cache correctly. The goal here is to do so.

First cache looked like this:

```
memoryCache
   [0] prouid=000-000-000-000
   [1] summaryVolumes
   [2] somethingElse
```

Invalidating the cache removed item at position 0, but the items at 1 and 2 remained and if they didn't time out would be fetched on the next request. Potentially being out of date.

The change now produces the following:

```
memoryCache
   [0] prouid=000-000-000-000
   [1] summaryVolumes
   [2] somethingElse

cacheKeys
   [0] prouid=000-000-000-000
       [0] summaryVolumes
       [1] somethingElse
```

Invalidating the cache uses the `subkeys` from `cacheKeys` to do a lookup and remove the actual response cache item from `memoryCache` (and `cacheKeys`). In this way only responses for the affected project are removed.

This implementation may not be 'ideal', but is widely considered the 'best' solution given the limitations we have with the provided memoryCache. Namely we cannot iterate the objects in memorycache.

Related work items: #68361