Merged PR 9051: Remove the calls to notify the Raptor service when a filter is created
_Cache in this instance refers to the Raptor proxy in-memory cache._

This follows closely behind changes @<David Glassenbury> made in commit `03ad016` to not notify Raptor if the filter is transient.

The 'notify raptor' calls on creation of a filter are redundant. The Raptor callback, `filterChange()`, removes the affected filter from cache which isn't necessary on a create, and also removes the project from the cache, also not necessary.

Changes in this PR address the following:
* Removed the calls to notify the Raptor service when a filter is created.
* Minor code tidying.

Related work items: #68808