Merged PR 9126: Fix design to ground tile request queries
We were incorrectly passing the filter in a design to ground query as the second Filter parameter to Raptor. In the scenario of either ground to design or design to ground this filter parameter should be the first filter param, the second filter param should be null. We were accidentally doing it correctly for ground to design. This story fixes design to ground requests.

I added a series of unit tests around the `TileRequestHelper` class that we didn't previously have.

There's general code tidying and namespace fixes too.

Related work items: #68588