Merged PR 9189: Add support for CTCT client and direct submission tag file request/responses
Versioned the current `/tagfiles` endpoint to satisfy the CTCT client requirements as per the design document linked at the bottom of this page.

Because of the client requirements a derived instance of the normally used `ContractExecutionResult` type is returned to the caller. Also the response status code is specifically set to 400 based on the value of the response object's `Code` property. See pages 9 and 10 of the design doc for response behaviours.

Beyond that there was minor code tidying of the Tag File controller, removed two unused source files, tidied the v2 Tag File endpoint executor. Because of the behaviour with the HTTP status code on the response I chose not to use the current v2 endpoint, and versioned it instead. The request object remains the same between the two.

Design Document: Earthworks Production Data Support
https://docs.google.com/document/d/1kTfi4LT-5uCdytCyW3YXQsyGMsls2thq6mfplENk47g/edit

Related work items: #65859