Merged PR 9407: Added Long to String converter to the MachineDetails::AssetId property.
The browser is not handling JSON responses that contain long properties with a value of 2^63. This fix resolves that issue by treating the MachineDetails::AssetId property as a string when serialized.

There's an argument to be made that I've coupled a client UI requirement to our model (because if you make a GET request outside of a browser the conversion of the long doesn't happen and therefore the string conversion is not necessary). However without a new endpoint we don't know if a request is from a browser or not.

Added other minor tidying/refactoring to the MachineDetails type.