Merged PR 9504: Remove unused package refs and DELETE sendMessage project
The other night I culled a number of package refs from the MD Consumer project that weren't being used. sendMessage also caught my eye and after looking around and talking to others it's apparently unused. Hence I've removed it from here. As you'll see there wasn't much too it.

Advice is that it was maybe used on an earlier version of Kafka. That version has gone and sendMessage along with it.