Merged PR 9600: Don't throw on tag file submission if Raptor returns 'invalid machine subscription'
**Current Behaviour**

If a TAG file is submitted via 3DP Raptor services and the exector doesn't succeed because of an invalid subscription error it throws an exception. You get 34 lines of exception details, but nothing useful or related to the TAG file in question.

Status Code: 400

```
{
    "Code": 2008,
    "Message": "Failed to process tagfile with error: OnChooseMachine. Machine Subscriptions Invalid."
}
```

**New Behaviour**

The result of the TAG file executor is not 'OK' and is 'tpsprOnChooseMachineInvalidSubscriptions' then we no longer throw an excption. The same JSON response and status code are returned to the user:

Status Code: 400

```
{
    "Code": 2008,
    "Message": "Failed to process tagfile with error: OnChooseMachine. Machine Subscriptions Invalid."
}
```

Ideally we'd return status code 200, but we still need to continue to return BadRequest for Tag File Harvester. A task should be added to update TFH to expect a 200 in the case of invalid subscription and instead interrogate the response object instead of relying on the status code.

With the changes made the 3DP Raptor services log file is less verbose, but more helpful:

2018-06-05 15:53:38,417 [3] INFO  RequestExecutorContainer req:e3174658-fb3c-49a3-a681-f3421aff46c0 {"Code":2008,"Message":"Failed to process tagfile with error: OnChooseMachine. Machine Subscriptions Invalid."}
2018-06-05 15:53:38,425 [3] INFO  RequestExecutorContainer req:e3174658-fb3c-49a3-a681-f3421aff46c0 TAG file submission request with file data ommitted:"{\"ProjectId\":1001158,\"FileName\":\"1112J010SW--B LHR934 S33251--120712065808.tag\",\"Boundary\":null,\"MachineId\":-1,\"TccOrgId\":null}"
2018-06-05 15:53:38,433 [3] INFO  ControllerActionInvoker req:e3174658-fb3c-49a3-a681-f3421aff46c0 Executed action method VSS.Productivity3D.WebApi.TagFileProcessing.Controllers.TagFileController.PostTagFile (VSS.Productivity3D.WebApi), returned result Microsoft.AspNetCore.Mvc.BadRequestObjectResult in 1225.6445ms.
...

Related work items: #69495