Merged PR 9599: Implementation of application service/cluster compute requests for profiling operations
It's a mixture of engine improvements and implementation of the request/response pattern for making profile requests.

It also includes an enhancement so surveyed surface patch requests to support composite elevation subgrid requests to be serviced in the same multi-surveyed surface topology logic as used for single valued surveyed surface requests.

Related work items: #71044