Merged PR 9735: Add new endpoint to return patches information using WorldOrigin rather than CellOrigin
Temporary solution to help CTCT team show internal demo by providing them with World Origin co-ords rather than cell origin.

Note, the only difference between the existing v1 endpoint and this new /worldorgin route is lines 111 and 112 in the `CompactionPatchExecutor` class. We need to calculate the value as meters instead of the bespoke cell orgin. This logic is a port of what Raptor does to achieve the same thing, (`SubGridTrees.pas, line 1520,  FIndexOriginOffset := 1 SHL (FNumBitsInKeys - 1); `)

This endpoint will be removed as part of the next story, #68810.

Related work items: #71477