Merged PR 9888: Fix 'dotnet build' error in TRexDesignElevationsServer and SurveyedSurfaceManager .csproj files
Attempting to `dotnet build` the Framework solution resulted in the following build errors:

`C:\Program Files\dotnet\sdk\1.0.4\Microsoft.Common.CurrentVersion.targets(2867,5): error MSB4216: Could not run the "GenerateResource" task because MSBuild could not create or connect to a task host with runtime "CLR4" and architecture "x86". Please ensure that (1) the requested runtime and/or architecture are available on the machine, and (2) that the required executable "C:\Program Files\dotnet\sdk\1.0.4\MSBuild.exe" exists and can be run.`

Consider this 'sustaining' work.
This commit fixes that problem and addresses minor tidying of the project files and R# meta files that were committed to the repo.

- Fixed 'dotnet build' error in TRexDesignElevationsServer and SurveyedSurfaceManager .csproj files.
- Removed unnecessary logging packages (duplicated because VSS.Log4NetExtensions.netcore is also included, and preferred).
- Removed .dotsettings files that are superseded by the .editorconfig file.