Merged PR 9901: Update Machine Details executor to use the new AS Node endpoint
The `GetOnMachineDesigns` endpoint has been reverted to the `Legacy` mode of behaviour. To restore 3DP to it's current and expected behaviour a new endpoint, `GetOnMachineDesignEvents()` endpoint has been added. This story is for the re-plumbing of that new endpoint.

This is to address a P0 bug in Alpha so needs to move through ASAP please.

Related work items: #72136