Merged PR 9989: Add boundary validation for tag file submissions
Problem: Tag files are being submitted that fall outside the project boundaries. This creates a 'data extents' problem in Raptor.

Manually submitted tag files required project Id. The service should have been resolving the boundary for the project and passing that on to Raptor. This has been fixed.

Automatically submitted tag files to not require a project Id; if no project Id is provided we do not send any boundary to Raptor.

Related work items: #72302