Merged PR 10039: Add Tpaas-Request-Id header value to the error log
Support issues between users/clients of the TPAAS endpoints may require the request id. To aid with future problem analysis this story is to put the logging of that header into the log if the request fails.

Affected lines are 229 and 295, other changes are the result of code tidying/formatting changes.

Related work items: #72678