Merged PR 10086: Reduce service controller DI complexity
This has been an outstanding MIP story for a number of months now. I found some time this evening so wanted to get it done.

Service controllers in 3DP have a large number of constructor DI services. Too many.

The goal here was to reduce controller complexity and improve testability by streamlining these service dependencies. A number of services were only used by the BaseController so these have been refactored to be lazy loaded and where possible removed from the derived service controllers.

A method from BaseController was also moved into an intermediate service aggregator class.

There is room for further tidying but we're getting diminished returns for the effort.