Merged PR 10188: Alter Patches cell Elevation to reduce overall cell size in the response
This piece of work is a follow up to the earlier #68810 where we provided a Patches endpoint for CTCT. This story alters the Elevation value returned from the endpoint from a float (4 bytes) to ushort (2 bytes), in an effort to reduce the overall response size sent back to the machine.

- The value returned by 3DP is also incremented by 1 for all non null values.
- Calculation of the elevation offset is now required to be done by the API consumer.

More work may follow as we further refine the Protobuf response behaviour.

Related work items: #68810