Merged PR 10248: Remove isLandfill checks from action filters
The Project ID and UID verifier action filters have shown to be a problem now that TBC is connecting directly to 3DP and not a legacy version of the service. On review it was decided the Landfill check once in place is no longer required so this has been removed.

Related work items: #73543