Merged PR 10263: DateRange information lost when Design or Alignment name are present
Following the earlier story to add DesignName and AlignmentName values to the FilterJSON this bug was found. It was caused because that new logic used the Filter::ToJsonString() method to serialize the Filter object. Apparently this method has fallen out of date and doesn't capture all properties on the type.

The fix was to use JSON.Net and serialize the object allowing the 'should serialize' properties on Filter to decide on the merit of serialization.

Added new unit tests for the name util class.

Related work items: #73632