Merged PR 10349: Make checking for archived projects simpler and improve on error message
This story deals with flipping the `ProjectVerifier` action filter's 'is archived' check. So it's now opt in. There is only one aspect of the application UI where archived projects should be rejected and that's TAG file submission.

The change here does that and improves the error message that's logged/returned so it's more clear why the request failed.

The problem was found after TBC switched from using a stand alone 3DP node to the 'current' production endpoint.

Related work items: #73813