Merged PR 10389: Add URL rewriter for malformed requests coming from TBC
TBC doesn't parse outbound request routes and through string concatenation of UI fields for host/endpoint values incorrect creates routes with occurrences of //, instead of the single forward slash.

When TBC was hitting the legacy 3DP endpoint (host recently deprecated) it was hosted under IIS which was doing automatic URL rewriting and so these paths weren't seen to cause an error. Since TBC has been updated to point to the new 3DP host we're seeing 404 errors in the client.

This story adds URL re writing to 3DP to parse out double forward slash occurrences in the path, replacing them with a single slash. There's an argument to be made that we could do this regardless of TBC, as a 'belts and braces' error resolver.

In lieu of this TBC needs to be fixed. Once done we can make the decision about whether the re writer in 3dp is needed.

Example of log activity when re writing occurs: (note the `v2//patches` request)

```
2018-08-15 10:04:02,616 [4] DEBUG Kestrel Connection id "0HLG267C4NFUE" stopped.
2018-08-15 10:04:02,713 [3] INFO  WebHost Request starting HTTP/1.1 GET http://localhost:5000/api/v2//patches?projectUid=ff91dd40-1569-4765-a2bc-014321f76ace&patchId=0&mode=0&patchSize=1&includeTimeOffsets=true
...
2018-08-15 10:04:02,932 [3] INFO  Startup req:a527bb11-88ea-4a06-9c4d-da640494e041 Path rewritten to: '/api/v2/patches'
```

Related work items: #73970