Merged PR 10369: Include optional time component in the Patches cell data
CTCT need the time of each patch cell in order to determine the most correct elevation value because they are maintaining and sharing elevation data between machines on the ground. To aid with this we now include an optional time delta in the cell object. It's summed with the timeOrigin value in the subgrid object to get the UTC time expressed as Unix seconds.

This story covers the addition of this calculation and a change to the patches v2 executor to handle a minified byte stream from Raptor.

Note the change to the `PatchExecutor` from reading DateTime to Double is correct. They both read 8 bytes, but semantically a Double is correct. Delphi expresses things differently and this had been carried over to 3DP. Now fixed.

Related work items: #72845