Merged PR 10392: Remove 3DP acceptance tests dependency on TEST_DATA_PATH env var
I've never been able to get AC tests working with this environment variable so I've removed it. There's now a helper that recursively searches up the directory tree from the executing assembly to locate the TestData folder. Works when running the tests from VS or via the command line `runacceptancetests.bat` against 3DP in the container.