Merged PR 10411: Added caching of legacy project id to the Raptor claims principal
Given a known project UID (Guid), the legacy Id (long) never changes. Though we repeatedly go and fetch it up to 3, maybe 4 times on some endpoints (including action filters). This change adds a simple <guid, long> cache in the RaptorPrincipal class.

It goes to the Proxy library which does maintain a limited cache of this stuff, this change just exits earlier and outlasts the purging of the proxy caches.

The static `ConcurrentDictionay<>` is thread safe.

It's a nice to have I did one evening, not married to it and not on the critical path.

Related work items: #74070