Merged PR 10448: Create new file upload endpoint that mimics the existing one without the FlowJs file wrapper
Created a new file upload endpoint that mimics the existing one without the FlowJs file wrapper but maintains file type and size restrictions.

The purpose of this endpoint is for non mobile users to submit files without the ceremony required by FlowJs.

It's known that the endpoint will receive very large TTM files (> 100mb, maybe larger) and so needed to support streaming of large files from the start.

The file upload logic itself is not very large, most of this PR is allowing the new `/direct` endpoint to piggyback off the existing controller acceptance tests.

If we wanted to drop FlowJs from the UI we could shift the UI to use this new endpoint, it would simply need to send a new `[string] filename` parameter on the request.

I expect we may come back and visit this once we start getting external users interacting with it, but in a sense that's by design.

I've also tested this using Postman and at 165mb TTM file. The response is exactly the same as the current endpoint.

```
{
    "importedFileDescriptor": {
        "projectUid": "fcd5eb47-8f83-4f9a-bd02-62d0f04f1850",
        "importedFileUid": "f875af1a-18b1-423c-80b8-a9d72f4871c3",
        "legacyFileId": 2000091,
        "customerUid": "be3a4693-1d97-48d8-af37-6db6a1974455",
        "importedFileType": 2,
        "dxfUnitsType": 0,
        "importedFileTypeName": "SurveyedSurface",
        "name": "56827_ob1c4bd5fd_1668_165mb.ttm",
        "fileCreatedUtc": "2018-03-27T11:00:00",
        "fileUpdatedUtc": "2018-04-01T12:00:00",
        "importedBy": "Utilization Develop CI",
        "surveyedUtc": "2017-09-11T00:00:00",
        "importedUtc": "2018-08-21T01:27:45.676844",
        "isActivated": true,
        "path": "/be3a4693-1d97-48d8-af37-6db6a1974455/fcd5eb47-8f83-4f9a-bd02-62d0f04f1850",
        "minZoomLevel": 0,
        "maxZoomLevel": 0,
        "importedFileHistory": [
            {
                "fileCreatedUtc": "2018-03-27T11:00:00",
                "fileUpdatedUtc": "2018-04-01T12:00:00"
            }
        ]
    },
    "Code": 0,
    "Message": "success"
}
```

Related work items: #66217