Merged PR 10748: Add DesignId to the /machinelifts response object
Added DesignId to the /machinelifts response object as per the story spec.

Prior to this change the objects in the `lifts` array didn't have the `designId` property.

```
{
  "MachineLiftsOneAsset": {
    "machineLiftDetails": [
      {
        "lifts": [
          {
            "designId": 1,
            "endUtc": "2014-04-10T03:51:06.274",
            "layerId": 8
          },
```

Note, I found we didn't have any existing tests for `machine lift ids`. We did have `lift ids` tests under a `machine lift ids` feature file so I renamed this in an earlier commit (outside this branch) and you'll see some of the 'fix up' of that in this commit. But we now have tests for `/machinelifts` and `/liftids` which are two distinct endpoints in 3DP.

Related work items: #72084