Merged PR 10749: Add logging to aid with future Report Controller failures
The goal of this story is add logging that will help with diagnosing problems in 3DP from requests made by the Report Service.

• Initially the `ReportController` had no logging; it now logs the request for each POST action method so we can more easily reproduce a failing request.
• Secondary was the unhandled exception throw in the `SummaryVolumesExecutor` because Raptor returns an error code not preloaded into the dynamically loaded `ContractExecutionStatesEnum`. This causes the Models package `GenericEnum.cs` class to throw an exception which in turn cancels out the ServiceException that's being thrown by the executor, meaning we lose that log message. I've guarded against this with a temporary code block and a reference to #74871 (Refactor Models::GenericEnum to not throw exceptions), which I'll address shortly in a separate piece of work. I think it's bad behaviour of the Models package GenericEnum class to throw back an exception.

Yes, the temporary try/catch I've added to `SummaryVolumesExecutor` is only on that one class and this problem has the potential to occur on any executor. That's the reason for opening story #74871, rather than refactor all executors in 3DP.

Related work items: #74074