Merged PR 10803: Add logging to help determine why logging isn't working
We are trying to debug a problem in the FileAccess service, but logging in the TCCFileAccess.netcore package doesn't work when deployed on a Linux environment.

From my local testing, I'm mildly optimistic that this will fix the problem. If not though, the additional logging may help.

• As part of the PR I've tidied up some of the local build/test scripts because this simplified my processes. These changes don't affect the actual CI build/deploy jobs as this FileAccess service is one of the new breed and those tasks are run out of the /build and /deploy folders.
• Part of my testing was to debug the running API while it was running in Docker. So you'll see some changes around the Docker support files for this reason.

There may be additional work to this story after we've deployed these changes and observed the results.

Related work items: #74729