Merged PR 10821: Change the content type from image/png to application/octet-stream
The `/rawfiles` endpoint was returning a content type `image/png` file when it wasn't returning an image, but a .ttm. Raptor didn't notice, doesn't care, because the result simply goes into it's cache and it's read from there, but for clarity I've fixed the logging and contentType of the endpoint to be correct. Note, I chose `application/octet-stream` because this subtype is the recommended content type for application specific file types, of which `.ttm` is.

Also:
• Addressed couple of code smells in the executor.
• Tweaked the logging to add clarity around what we're seeing.
• Minor tidying of the `FileDescriptor` class.

Related work items: #75170