Merged PR 10839: Fix ProjectVerifier
Commit `7e222cda` (17 Aug '18) introduced a bug in the TagFileController class, forcing project id validation on all endpoints of the controller, not just a few. Included in this was the `/direct` TAG file submission endpoint. It's expected that requests to this endpoint would NOT provide a project Id, so this check needed to be undone.

During investigation and testing of this fix it was apparent we where doing some things twice, or wrong, or not at all with respect to project id validation.

The following changes work to correct these, they are:

• Fixed some POST methods that named the input parameter anything but 'request'. This is because `ProjectVerifier` checks for the `request` parameter. So in these cases they were not correctly validating project ids. (A future task might be to make the `ProjectVerifier` less pedantic about what `[FromBody]` parameter to look for.
• Removed the `isArchived` check from `ProjectVerifier` as it was only used by one method (tag file upload), to where I moved the related check.
• Removed the `ProjectIdJsonInputFormatter` class. It's role as request middleware was to normalise all requests, ensuring that regardless of whether the legacy `ProjectId` was set, or the new `ProjectUid`, by the time the request got to the controller, it would be the `ProjectId` that was populated. I think the thinking here was to cut down on id checking work at the controller level. This behaviour was effectively being done twice (because of the `ProjectVerifier` action filter). It didn't cater for ProjectId/Uid values in the request body, which is why the ProjectVerifier action filter was created. This change encapsulates all project id checking in the one action filter, and removes middleware from the request processing.
• Fixed the `ProjectVerifier` to correctly check inherited properties on the request object when looking for `ProjectId` or `ProjectUid`. It wasn't doing this before so request DTOs that inherited from `ProjectID.cs` were not being correctly vetted.
• Fixes to the `ProjectVerifier` showed we had some acceptance tests checking for the wrong response code. e.g. They expected Unauthorized for a bad or null project id when it should have been BadRequest. These came to light following the above changes and have been fixed.

Now, neither the `ProjectIdJsonInputFormatter` middleware or `ProjectVerifier` make requests via `RaptorPrincipal` to the project service to get either the legacy id or to check the `isArchived` property so we are now doing much less work during the id parsing phase for each request.

Related work items: #75173