Merged PR 10827: Change GenericEnum::FirstNameWith() to not throw if the error code isn't found
Changed GenericEnum::FirstNameWith() to not throw if the error code isn't found.

When these occur they can hide the real error that's attempting to be reported on.

e.g. take the following:

```
        throw new ServiceException(HttpStatusCode.BadRequest, new ContractExecutionResult((int) raptorResult,
            $"Failed to get requested volumes summary data with error: {ContractExecutionStates.FirstNameWithOffset((int)raptorResult)}"));
```

If the `raptorResult` value isn't found in the dynamically loaded enum `GenericEnum::FirstNameWith()` will throw. This means the attempt to throw the ServiceException is not completed meaning we lose contextual logging of why the Raptor request failed.

This change means the application behaviour doesn't change if the code isn't found, we just get the 'real' exception and an error message letting us know the code wasn't found.

Related work items: #74870