Merged PR 10862: #75288 - Modify Scheduler to allow binary data in the schedule job request
Modify Scheduler to allow binary data in the schedule job request (previously it was all strings, which caused encoding / decoding issues).

Also modified Scheduler to store the actual request in S3, as opposed to the database - as the Request holds the entire POST body (files) uploaded, which makes the DB grow quite quickly.

***NOTE: This version of Scheduler will not be backwards compatible with any Jobs already scheduled in the Hangfire DB between the older versions and this version. It will however be backwards compatible with any web requests to create new jobs.***

Related work items: #75288