Merged PR 10955: Update controller to return correct NoContent results, without trying to set the response body
Commit `94db6fe9` changed the behaviour of the `CompactionSummaryDataController` so it no longer returned a BadRequest code if there was no content returned from Raptor. The change was to return NoContent instead. Unfortunately it leverages the `ServiceException` class which tried to set the response `Message` value to that of the executors error message. In the application logs at runtime this produces the following framework exception:

System.InvalidOperationException: Writing to the response body is invalid for responses with status code 204.
   at Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http.Frame.HandleNonBodyResponseWrite()

The result is largely the same for the UI, as they're only checking for 200 or !200, but it was creating quite a mess in the application logs.

With the changes made in this PR there are now no exceptions thrown in the controller. Logging for BadRequest responses is therefore done using a direct call to logger.

Related work items: #74743