Merged PR 11171: Minor updates to the DockerFile_win file
As a developer I want to refactor the dockerfile_win file and move away from using the build pattern, and instead use more layers so the work to download and install the .NET framework doesn't get repeated. We can leverage caching more through this multi-stage approach.

https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#leverage-build-cache

The old (builder) pattern here was popular in earlier versions of Docker where more layers was bad, this is (typically) no longer the case.

Related work items: #76329