Merged PR 11320: #75997 -Add center point to the metadata to allow for placement on the 2d map
To place the 3d mesh correctly on the 2d map we need to know the center point of the mesh (as well as the extents).

Without it we can't tell where each vertex is meant to be position, as they're are the center in the resulting obj file.

Related work items: #75997