Merged PR 11331: #76608 - Stop BaseProxy from caching failed results by only caching valid
Previously the Cache used by BaseProxy would cache results even if they were failed results (such as a contract exception).
Then the next request coming in would hit the cache, and return the same result, an error - even if the error was no longer valid.

To fix this, the Cache is now configured to only return successful results (i.e anything that didn't throw an exception). This will fix the majority of caching issues with BaseProxy.

A test to confirm it's working is to hit an endpoint that uses BaseProxy with an invalid Bearer token, and get an 'Authorization failure' error, then attempt the same request with a valid Bearer token. Previously these were cached, causing the error to be returned again. This is no longer the case.

Note: 3dp and other services will need to update their proxy package to see the changes

Related work items: #76608