Merged PR 11397: Code tidying and refactoring
The 2076 error code not found problem reported in this story was fixed in 18 September 2018 in the MasterDataModels package, commit id `31f26c6`.

During investigation of this I noticed some code that could be tidied, so that's what we're addressing here;

1. Changed project to not treat package downgrade warnings as errors.
2. Reverted 3DP WebApi Models project to the earlier namespace (put the period back).
3. Refactored Summary Volumes exectors, moving common result conversion logic into a helper class; added unit test.
4. Pulled in project reference to MockWebAPI so if you now set the solution to build multiple projects you'll get 3DP WebAPI and MockWebAPI both running on one keystroke! Finally.

The story this task is attached too will remain open while we investigate other related errors on the POST /v1/summaryvolumes endpoint.

Related work items: #75371