Merged PR 11603: Add global solution file for the mono repository
As a developer I need the ability to easily consolidate packages across projects in the mono repository.

You don't really need to review this; it's simply a solution file for all projects sans TRex and updates and consolidation of some packages so that the solution will be allowed to build.

If someone wanted before creating a PR for Proxies, or 3DP they could try building this solution file to ensure there weren't breaking changes introduced outside of where they were visibly working.

Related work items: #77058