Merged PR 11636: Fix Coordinate Service unit tests to use new TPaaSClient DI setup pattern
Coordinate service unit tests were broken following recent changes to the way the TPaaSClient DI process is done. This story addresses those changes in the `CoordinatesAPIClientTestDIFixture` class.

* You'll see other minor code tidying.
* The new PowerShell script to set environment variables is akin to what we use in 3DP for setting vars locally for use when running locally (and not inside docker/helm, etc. You'll see in this case it works in with the unit test setup.

It's acknowledged that coordinate service DI that's in use outside of the unit tests, e.g. Filter Utilities or Profiling is not working yet. That is outside the scope of this story and will be resolved soon...

Related work items: #77404